refactor: 패키지 이동 및 헤더 파싱 로직 이동
