+++ tomcat/src/main/java/org/apache/coyote/http11/HttpProtocolVersion.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.reqeust;
+package org.apache.coyote.http11;
 
 import java.util.Arrays;
