+++ tomcat/src/main/java/org/apache/coyote/http11/reqeust/HttpRequest.java
@@ -1,5 +1,8 @@
 package org.apache.coyote.http11.reqeust;
 
+import org.apache.coyote.http11.HttpHeaders;
+import org.apache.coyote.http11.HttpProtocolVersion;
+
 public record HttpRequest(
         HttpMethod method,
         String uri,
