+++ tomcat/src/main/java/org/apache/coyote/http11/reqeust/util/HttpRequestParser.java
@@ -1,9 +1,7 @@
 package org.apache.coyote.http11.reqeust.util;
 
-import java.util.List;
-import org.apache.coyote.http11.reqeust.HttpHeaders;
+import org.apache.coyote.http11.HttpProtocolVersion;
 import org.apache.coyote.http11.reqeust.HttpMethod;
-import org.apache.coyote.http11.reqeust.HttpProtocolVersion;
 
 public class HttpRequestParser {
 
@@ -26,16 +24,6 @@ public class HttpRequestParser {
         final String httpVersion = startLine.split(" ")[2];
 
         return HttpProtocolVersion.valueOfVersion(httpVersion);
-    }
-
-    public HttpHeaders parseHeaders(final List<String> headerLines) {
-        final HttpHeaders headers = new HttpHeaders();
-        for (final String rawHeader : headerLines) {
-            final String[] rawHeaders = rawHeader.split(": ");
-            headers.addHeader(rawHeaders[0], rawHeaders[1]);
-        }
-
-        return headers;
     }
 
     public static HttpRequestParser getInstance() {
